+++ config/crd/bases/gmos.chtc.wisc.edu_glideinsets.yaml
@@ -36,6 +36,9 @@ spec:
             description: Collection of Glideins sharing a priority class, per-pod
               resource allocation, and node affinity
             properties:
+              name:
+                description: Name of this glidein deployment
+                type: string
               nodeSelector:
                 additionalProperties:
                   type: string
@@ -136,6 +139,8 @@ spec:
                       type: string
                   type: object
                 type: array
+            required:
+            - name
             type: object
           status:
             description: GlideinManagerPilotSetStatus defines the observed state of
