+++ internal/controller/glideinmanagerpilotset_controller.go
@@ -131,6 +131,9 @@ func (r *GlideinManagerPilotSetReconciler) Reconcile(ctx context.Context, req ct
 		return ctrl.Result{}, nil
 	}
 
+	if err := CreateResourcesForPilotSet(r, ctx, pilotSet); err != nil {
+		return ctrl.Result{}, err
+	}
 	psState := &PilotSetReconcileState{reconciler: r, ctx: ctx, resource: pilotSet}
 	AddGlideinManagerWatcher(pilotSet, psState)
 	AddCollectorClient(pilotSet, psState)
@@ -214,76 +217,10 @@ func (pu *PilotSetReconcileState) ApplySecretUpdate(secSource PilotSetSecretSour
 	return ApplyUpdateToResource(pu, RNTokens, &corev1.Secret{}, &TokenSecretValueUpdater{secSource: &secSource, secValue: &sv})
 }
 
-// Create a new Kubernetes object:
-// 1. Check that a resource with the given name doesn't yet exist in the namespace
-// 2. Create an initial schema for the object in-memory
-// 3. Post the newly created object to k8s via the API
-func CreateResourceIfNotExists[T client.Object](reconcileState *PilotSetReconcileState, resourceName ResourceName, resource T, creator ResourceCreator[T]) error {
-	log := log.FromContext(reconcileState.ctx)
-	name := resourceName.NameFor(reconcileState.resource)
-	if err := reconcileState.reconciler.GetClient().Get(
-		reconcileState.ctx, types.NamespacedName{Name: name, Namespace: reconcileState.resource.GetNamespace()}, resource); err == nil {
-		log.Info("Resource already exists, no action needed.")
-	} else if apierrors.IsNotFound(err) {
-		log.Info("Resource not found, creating it.")
-		resource.SetName(name)
-		resource.SetNamespace(reconcileState.resource.GetNamespace())
-		if err := creator.SetResourceValue(reconcileState.reconciler, reconcileState.resource, resource); err != nil {
-			log.Error(err, "Unable to set value for new resource")
-		}
-		if err := ctrl.SetControllerReference(reconcileState.resource, resource, reconcileState.reconciler.GetScheme()); err != nil {
-			return err
-		}
-		if err := reconcileState.reconciler.GetClient().Create(reconcileState.ctx, resource); err != nil {
-			log.Error(err, "Unable to create resource")
-			return err
-		}
-		MarkNamespaceOutOfSync(reconcileState.resource.GetNamespace())
-		return nil
-	} else {
-		log.Error(err, "Unable to get resource")
-		return err
-	}
-	return nil
-}
-
-// Update an existing Kubernetes object:
-// 1. Fetch the object by name via the k8s API
-// 2. Modify the object's data in-memory
-// 3. Push the updated data back to k8s via the API
-func ApplyUpdateToResource[T client.Object](reconcileState *PilotSetReconcileState, resourceName ResourceName, resource T, resourceUpdater ResourceUpdater[T]) error {
-	log := log.FromContext(reconcileState.ctx)
-	name := resourceName.NameFor(reconcileState.resource)
-	if err := reconcileState.reconciler.GetClient().Get(
-		reconcileState.ctx, types.NamespacedName{Name: name, Namespace: reconcileState.resource.GetNamespace()}, resource); err == nil {
-		updated, err := resourceUpdater.UpdateResourceValue(reconcileState.reconciler, resource)
-		if err != nil {
-			log.Error(err, "Unable to apply update to resource value")
-			return err
-		}
-		if !updated {
-			log.Info("No updates needed for resource")
-			return nil
-		}
-		if err := reconcileState.reconciler.GetClient().Update(reconcileState.ctx, resource); err != nil {
-			log.Error(err, "Unable to post update to resource")
-			return err
-		}
-		log.Info("Resource updated successfully")
-	} else if apierrors.IsNotFound(err) {
-		log.Info("Resource not found, must have been deleted or not created")
-		return err
-	} else {
-		log.Error(err, "Unable to get resource")
-		return err
-	}
-	return nil
-}
-
 func CreateResourcesForPilotSet(r *GlideinManagerPilotSetReconciler, ctx context.Context, pilotSet *gmosv1alpha1.GlideinManagerPilotSet) error {
 	log := log.FromContext(ctx)
 	log.Info("Got new value for PilotSet custom resource!")
-	psState := &PilotSetReconcileState{reconciler: r, ctx: ctx, resource: &gmosv1alpha1.GlideinSet{}}
+	psState := &PilotSetReconcileState{reconciler: r, ctx: ctx, resource: pilotSet}
 
 	// Collector resources
 	log.Info("Creating Collector Signing Key if not exists")
@@ -306,37 +243,14 @@ func CreateResourcesForPilotSet(r *GlideinManagerPilotSetReconciler, ctx context
 		return err
 	}
 
-	return nil
-}
-
-func CreateResourcesForGlideinSet(r *GlideinSetReconciler, ctx context.Context, pilotSet *gmosv1alpha1.GlideinSet) error {
-	log := log.FromContext(ctx)
-	log.Info("Got new value for GlideinSet custom resource!")
-	psState := &PilotSetReconcileState{reconciler: r, ctx: ctx, resource: pilotSet}
-
-	log.Info("Creating Tokens secret if not exists")
-	if err := CreateResourceIfNotExists(psState, RNGlideinTokens, &corev1.Secret{}, &EmptySecretCreator{}); err != nil {
-		return err
-	}
-
-	// PilotSet
-	log.Info("Creating Data Secret if not exists")
-	if err := CreateResourceIfNotExists(psState, RNData, &corev1.Secret{}, &EmptySecretCreator{}); err != nil {
-		return err
-	}
-
-	log.Info("Creating Access Token Secret if not exists")
-	if err := CreateResourceIfNotExists(psState, RNTokens, &corev1.Secret{}, &EmptySecretCreator{}); err != nil {
-		return err
-	}
-
-	log.Info("Creating Deployment if not exists")
-	if err := CreateResourceIfNotExists(psState, RNBase, &appsv1.Deployment{}, &PilotSetDeploymentCreator{}); err != nil {
+	log.Info("Creating GlideinSet[0] if not exists")
+	glideinSetSpec := pilotSet.Spec.GlideinSets[0]
+	gsResource := ResourceName("-" + glideinSetSpec.Name)
+	if err := CreateResourceIfNotExists(psState, gsResource, &gmosv1alpha1.GlideinSet{}, &GlideinSetCreator{spec: &glideinSetSpec}); err != nil {
 		return err
 	}
 
 	return nil
-
 }
 
 func updateErrOk(err error) bool {
